problem: log logic is clumped awkwardly
solution: refactor to files
