+++ cmd/geth/log_display_util.go
@@ -1,0 +1,47 @@
+package main
+
+import (
+	"time"
+	"gopkg.in/urfave/cli.v1"
+	"github.com/ethereumproject/go-ethereum/eth"
+	"github.com/ethereumproject/go-ethereum/eth/downloader"
+)
+
+var lsModeName = []string{
+	"Discover",
+	"Sync",
+	"Fast",
+	"Import",
+}
+
+// Global bookmark vars.
+// These are accessible globally to allow inter-communication between display system event handlers.
+// TODO: ensure handler cooperation; ie use a mutex, atomic, or something
+var currentMode = lsModeDiscover
+var currentBlockNumber uint64
+var chainEventLastSent time.Time
+
+// updateLogStatusModeHandler implements the displayEventHandlerFn signature interface
+// It is a convenience fn to update the global 'currentMode' var.
+// Typically it should be called from downloader events, and uses the 'getLogStatusMode' logic.
+func updateLogStatusModeHandler(ctx *cli.Context, e *eth.Ethereum, evData interface{}, tickerInterval time.Duration) {
+	currentMode = getLogStatusMode(e)
+}
+
+// getLogStatusMode gets the "mode" for the ethereum node at any given time.
+// It is used to set the global bookmark variable, and influences formatting logic.
+func getLogStatusMode(e *eth.Ethereum) lsMode {
+	if e.Downloader().Synchronising() {
+		switch e.Downloader().GetMode() {
+		case downloader.FullSync:
+			return lsModeFullSync
+		case downloader.FastSync:
+			return lsModeFastSync
+		}
+	}
+	_, current, height, _, _ := e.Downloader().Progress() // origin, current, height, pulled, known
+	if e.Downloader().GetPeers().Len() > 0 && current >= height && !(current == 0 && height == 0) {
+		return lsModeImport
+	}
+	return lsModeDiscover
+}
